加入按键长按连加， 写入数据到内部flash程序（未完成）

+++ PulseFenceManage/Bsp/master_manage.h
@@ -51,8 +51,8 @@ typedef enum{
 typedef struct
 {
 	zone_type_def zone_type;
-	uint16_t zone1_id;
-	uint16_t zone2_id;
+	uint8_t zone1_id;
+	uint8_t zone2_id;
 	zone_mode_def zone_mode;
 	zone_voltage_level_def zone_voltage_level;
 	zone_status_def zone1_sta;
